simplified state-level parsing via the GAE API, only one script now
